api/requirements.txt
flask
pandas
numpy
scikit-learn
joblib
gunicorn
psutil
pyyaml